Move input-related elements to the 'input' element group folder
