+++ angular/components/elements/element-group-input/dropdown/dropdown.component.html
