+++ angular/components/elements/element-group-input/dropdown/dropdown.component.scss
@@ -1,8 +1,8 @@
 // dropdown.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 dropdown {
     @include flex(center, center, row);
