+++ angular/components/elements/element-group-input/input-popup/input-popup.component.scss
@@ -1,0 +1,9 @@
+// input-popup.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+input-popup {
+
+}
