+++ angular/components/elements/element-group-input/input-popup/input-popup.component.ts
@@ -1,8 +1,8 @@
 // input-popup.component.ts 
 import {Component, Input} from "@angular/core";
-import {TagType} from "../../../models/html/TagType";
-import {BillInputPopupType} from "../element-group-bill/bill-input/models/BillInputPopupType";
-import {IncomeInputPopupType} from "../element-group-income/income-input/models/IncomeInputPopupType";
+import {TagType} from "../../../../models/html/TagType";
+import {BillInputPopupType} from "../../element-group-bill/bill-input/models/BillInputPopupType";
+import {IncomeInputPopupType} from "../../element-group-income/income-input/models/IncomeInputPopupType";
 
 @Component({
     selector: 'input-popup',
