+++ angular/components/elements/element-group-input/input-time-type-selector/input-time-type-selector.component.html
