+++ angular/components/elements/element-group-input/input-time-type-selector/input-time-type-selector.component.scss
@@ -1,8 +1,8 @@
 // input-time-type-selector.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 input-time-type-selector {
     @include flex(flex-start, center, row);
