+++ angular/components/elements/element-group-input/name-input/name-input.component.scss
@@ -1,0 +1,9 @@
+// name-input.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+name-input {
+
+}
