+++ angular/components/elements/element-group-input/work-input/work-input.component.scss
@@ -1,0 +1,9 @@
+// work-input.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+work-input {
+
+}
