--- angular/components/elements/input-popup/input-popup.component.scss
@@ -1,9 +1,0 @@
-// input-popup.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-input-popup {
-
-}
