--- angular/components/elements/work-input/work-input.component.scss
@@ -1,9 +1,0 @@
-// work-input.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-work-input {
-
-}
